separate the pcb svg export
